refactor: split osty CLI native check and lint helpers into files
Made-with: Cursor
